add utils for viewer
